Remove godeps
Running godeps didn't work for us, because we depended on generated
resources.  While it did provide functionality around locking down
versions, we manually specify the CF version that we depend on, and this
is the main dependency of the project.  If it needs to be introduced
again, it should be added to the deps script.