Allow go get to work from branches.
No longer use -u to get deps from the network. Go get -u defaults
to master branch with no way of pointing to a branch other than
explicitly setting them in the import statement. We originally needed
this so that incrementing the CF_CLI_VERSION would allow the checkout
to succeed as it has the latest code. We now do an explicit checkout.

This change will now mean that dependencies locally will not necessarily
be the latest. They will however be the latest in the CI.