Support Go 1.7 and CLI version 6.22.1
This commit updates the bgd plugin to compile with go 1.7
following the suggestions in issue #13 and fixes the
**Failed to get default shared domain** error reported in issue #11.

The **Failed to get default shared domain** error was happening because
of a change in the behaviour of
`CfPlugin.Connection.CliCommandWithoutTerminalOutput`
which used to return a `[]string` with the entire contents of the
response in index 0. It now returns the output of the CliCommand split
by "\n". So we need to join it again before trying to unmarshall it.

DCO 1.1 Signed-off-by: Ravi Kotecha <kotecha.ravi@gmail.com>